update for page reload
its working but there is one small issue with this code. when user mouse is in music iframe till then only it doesn't reload the page. but when user mouse is in outside of music iframe and song is playing then it reloads the page. but it shouldn't reload the page when user mouse is outside of music iframe and music is playing i.e. if music iframe is playing music in background then it shouldn't reload the page.